Refactored ReplayWorkflowExecutor
