+++ src/main/java/io/temporal/internal/replay/StatefulTaskHandler.java
@@ -22,10 +22,15 @@ package io.temporal.internal.replay;
 import io.temporal.api.common.v1.Payloads;
 import io.temporal.api.query.v1.WorkflowQuery;
 import io.temporal.api.workflowservice.v1.PollWorkflowTaskQueueResponseOrBuilder;
-import java.time.Duration;
 import java.util.Optional;
 
-public interface WorkflowExecutor {
+/**
+ * Task handler that encapsulates a cached workflow and can handle multiple calls to
+ * handleWorkflowTask for the same workflow run.
+ *
+ * <p>Instances of this object can be cached in between workflow tasks.
+ */
+public interface StatefulTaskHandler {
 
   /**
    * Handles a single workflow task.
@@ -40,6 +45,4 @@ public interface WorkflowExecutor {
       PollWorkflowTaskQueueResponseOrBuilder workflowTask, WorkflowQuery query);
 
   void close();
-
-  Duration getWorkflowTaskTimeout();
 }
