+++ src/main/java/io/temporal/internal/statemachines/WorkflowTaskStateMachine.puml
@@ -26,6 +26,7 @@ title WorkflowTask State Transitions
 [*] --> CREATED
 CREATED --> SCHEDULED: WORKFLOW_TASK_SCHEDULED
 SCHEDULED --> STARTED: WORKFLOW_TASK_STARTED
+SCHEDULED --> TIMED_OUT: WORKFLOW_TASK_TIMED_OUT
 STARTED --> COMPLETED: WORKFLOW_TASK_COMPLETED
 STARTED --> FAILED: WORKFLOW_TASK_FAILED
 STARTED --> TIMED_OUT: WORKFLOW_TASK_TIMED_OUT
