+++ src/test/java/io/temporal/internal/replay/ReplayWorkflowStatefulTaskHandlerTaskHandlerTests.java
@@ -43,7 +43,7 @@ import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 
-public class ReplayWorkflowExecutorTaskHandlerTests {
+public class ReplayWorkflowStatefulTaskHandlerTaskHandlerTests {
 
   private TestWorkflowService testService;
   private WorkflowServiceStubs service;
